+++ .changeset/actionmenu-remove-focus-trap.md
@@ -2,4 +2,4 @@
 '@primer/react': patch
 ---
 
-ActionMenu: Remove focus trap to enable Tab and Shift+Tab behavior
+# ActionMenu: Remove focus trap to enable Tab and Shift+Tab behavior
