+++ .changeset/chatty-moose-wait.md
@@ -2,4 +2,4 @@
 '@primer/react': minor
 ---
 
-Adds the option to render a trailing action inside of the TextInput component
+# Adds the option to render a trailing action inside of the TextInput component
