+++ .changeset/config.json
@@ -1,10 +1,15 @@
 {
-  "$schema": "https://unpkg.com/@changesets/config@1.4.0/schema.json",
-  "changelog": ["@changesets/changelog-github", {"repo": "primer/react"}],
-  "commit": false,
-  "linked": [],
-  "access": "public",
-  "baseBranch": "main",
-  "updateInternalDependencies": "patch",
-  "ignore": []
+	"$schema": "https://unpkg.com/@changesets/config@1.4.0/schema.json",
+	"changelog": [
+		"@changesets/changelog-github",
+		{
+			"repo": "primer/react"
+		}
+	],
+	"commit": true,
+	"linked": [],
+	"access": "public",
+	"baseBranch": "main",
+	"updateInternalDependencies": "patch",
+	"ignore": []
 }
