+++ .changeset/six-trains-guess.md
@@ -2,4 +2,4 @@
 '@primer/react': patch
 ---
 
-Updates the API for token components to align with our size-naming ADR, avatar guidelines, and icon guidelines
+# Updates the API for token components to align with our size-naming ADR, avatar guidelines, and icon guidelines
