+++ .changeset/spicy-carrots-remember.md
@@ -2,4 +2,4 @@
 "@primer/react": patch
 ---
 
-Update `@primer/behaviors` for better ESM compatibility
+# Update `@primer/behaviors` for better ESM compatibility
