+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_SERVER_URL=https://example.com/api/
+REACT_APP_CLIENT_ID=1097hy67546hbdy2
