+++ README.md
@@ -1,6 +1,6 @@
 # Folder Structure Conventions  
 
-# <Folder structure options and naming conventions for software projects>
+# Folder structure options and naming conventions for software projects
 
 A typical top-level directory layout
 .
