+++ package.json
@@ -1,26 +1,33 @@
 {
   "name": "axtarilanproqramcilar",
-  "version": "0.1.0",
+  "version": "3.0.0",
   "private": true,
   "dependencies": {
     "@primer/css": "^19.8.2",
+    "@react-three/drei": "9.1.5",
+    "@react-three/fiber": "8.0.10",
     "@reduxjs/toolkit": "^1.8.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^14.0.4",
     "base-js": "^1.0.1",
+    "bootstrap": "5.1.3",
     "cjs": "^0.0.11",
     "http-proxy-middleware": "^2.0.4",
     "nan": "^2.15.0",
-    "react": "^18.0.0",
+    "react": "18.0.0",
     "react-dom": "^18.0.0",
     "react-helmet": "^6.1.0",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.8",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.0",
+    "react-terminal": "1.2.1",
+    "reactstrap": "9.0.1",
     "select2": "4.1.0-rc.0",
-    "styled-components": "^5.3.5"
+    "styled-components": "5.3.5",
+    "three": "0.139.2",
+    "use-animated-sprite": "1.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
