+++ src/features/extracomponents/Customers/UserProfile.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UserProfile = () => {
+  return (
+    <div></div>
+  )
+}
+
+export default UserProfile
