+++ src/features/extracomponents/Employees/Freelancers.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Freelancers = () => {
+  return (
+    <div></div>
+  )
+}
+
+export default Freelancers
