+++ src/features/extracomponents/JoinLogic/Home.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div></div>
+  )
+}
+
+export default Home
