+++ src/features/root/Dashboard.js
@@ -1,21 +1,21 @@
 /* eslint-disable no-unused-vars */
 // @ts-nocheck
-import React from "react";
+import React, { useState } from "react";
 import { Routes, Route } from "react-router-dom";
-import PasswordRecovery from "../common/passwordrecovery/PasswordRecovery";
-import Register from "../common/register/Register";
-import Login from "../common/login/Login";
+import Register from "./../common/register/Register";
+import Login from "./../common/login/Login";
 
 const Dashboard = () => {
-  return (
-    <div className="commonContainer">
-      <div className="navTop">
-          <div className="navBar"></div>
-          <div className="aside"></div>
-          <div className="main"></div>
-          <div className="footer"></div>
-      </div>
+  const [customer, setCustomer] = useState(false);
+  const [freelancer, setFreelancers] = useState(false);
 
+  return (
+    <div>
+      <Routes>
+          <Route path="/" element={<Register customer={customer} freelancer={freelancer} />} />
+          <Route path="/login" element={<Login customer={customer} freelancer={freelancer} />} />
+
+      </Routes>
     </div>
   );
 };
