+++ src/index.js
@@ -7,6 +7,7 @@ import { BrowserRouter } from "react-router-dom";
 import store from "./app/store";
 import App from "./features/root/App";
 import "./index.css";
+import "bootstrap/dist/css/bootstrap.min.css"
 
 import inject from "./inject";
